will no longer save data to disk if nothing has changed
notepad++ kept asking me to reload the file even when nothing had changed, this prevents that annoyance

+++ PENIS For Unity/DataFile.cs
@@ -58,7 +58,10 @@ namespace PENIS
         public void SaveAllData()
         {
             string PENIS = DataConverter.PENISFromDataStructure(TopLevelLines);
-            File.WriteAllText(FilePath, PENIS);
+            string ExistingPENIS = File.ReadAllText(FilePath);
+
+            if(PENIS != ExistingPENIS)
+                File.WriteAllText(FilePath, PENIS);
         }
